[MODIFY] Update the PD deck code to be in uin16_t and to use the CF SPI API.

+++ src/deck/drivers/interface/photodiode_deck.h
@@ -42,7 +42,7 @@
  * Thread-safe (mutex-protected). Max blocking: 5 ms.
  * Returns false if deck not ready or mutex timeout.
  */
-bool pdDeckGetValues(float out[PD_CHANNEL_COUNT]);
+bool pdDeckGetValues(uint16_t out[PD_CHANNEL_COUNT]);
 
 /** Returns true once at least one conversion has completed. */
 bool pdDeckIsReady(void);
